Build: Update dependency analysis configuration on app
These 'unused' related advises for the 'app' module is actually
correct, both 'media-placeholders' and ':picasso-loader', can be removed
as dependencies since they are actually unused and not being utilized
from within the app.

However, I chose to keep them there and suppress the advise instead in
order to preserve the app default configuration, its purpose being to
test all those individual modules:
- aztec
- glide-loader
- picasso-loader
- wordpress-comments
- wordpress-shortcodes
- media-placeholders

As such this advice is instead suppressed and the 'onUnusedDependencies'
exclusion configuration were used to achieve that.

------------------------------------------------------------------------

Advice for :app
Unused dependencies which should be removed:
  implementation project(':media-placeholders')
  implementation project(':picasso-loader')