Deps: Replace androidx legacy support v4 dependency in aztec
Although removing the 'androidx.legacy.support.v4' is enough by itself
and the build is (almost) not failing, it is better to also replace that
with 'androidx.collection' due to this 'androidx.collection.ArrayMap'
import on 'AztecQuoteSpan'.

FYI: It got introduced as part of this
(ad68cc9b640e7f50d7e21e7e459083b45fa1a7d7) commit, a part of the
AndroidX migration.

PS: To make sure that the build is not failing the only instance of
'androidx.legacy', the '<androidx.legacy.widget.Space/>' got replaced
with '<Space/>' on the 'aztec_format_bar_advanced.xml' view.

------------------------------------------------------------------------

This replacement was suggested by the dependency analysis report,
see below:

Advice for :aztec
Unused dependencies which should be removed:
  implementation 'androidx.legacy:legacy-support-v4:1.0.0'

These transitive dependencies should be declared directly:
  ...
  implementation 'androidx.collection:collection:1.1.0'
  ...
  implementation 'androidx.legacy:legacy-support-core-ui:1.0.0'
  ...