Build: Update dependency analysis configuration on media placeholders
This 'unused' related advise for the 'media-placeholders' module is
actually incorrect as the build is failing when the 'appcompat'
dependency is removed.

This is happening due to the fact that 'aztec' has classes like
'AztecText', which is being used on this module, but the Dependency
Analysis plugin cannot understand that although there are no 'appcompat'
related imports, those are actually needed nevertheless. Another
dependency configuration might be able to solve this, but this was
considered as being outside the scope of this change.

------------------------------------------------------------------------

Advice for :media-placeholders
Unused dependencies which should be removed:
  implementation 'androidx.appcompat:appcompat:1.1.0'