Build: Update dependency analysis configuration on wordpress shortcodes
This 'unused' related advise for the 'wordpress-shortcodes' module is
actually incorrect as the build is failing when the 'appcompat'
dependency is removed.

Just like it was done on this (7db2a43daf2da3de032dd5f063c62c33a114dca1)
change, the seemingly unused 'androidx.appcompat' dependency advise got
suppressed as well.

------------------------------------------------------------------------

Advice for :wordpress-shortcodes
Unused dependencies which should be removed:
  implementation 'androidx.appcompat:appcompat:1.1.0'